requirements.txt
asgiref==3.8.1
Django==5.1.4
django-ckeditor==6.7.1
django-cron==0.6.0
django-js-asset==2.2.0
djangorestframework==3.15.2
mysqlclient==2.2.6
pillow==10.4.0
psutil==6.1.0
python-decouple==3.8
sqlparse==0.5.2
tzdata==2024.1
